fix: re-change id type from Long to Int

+++ server/src/main/kotlin/it/polito/wa2/g29/server/dto/ProductDTO.kt
@@ -3,7 +3,7 @@ package it.polito.wa2.g29.server.dto
 import it.polito.wa2.g29.server.model.Product
 
 data class ProductDTO(
-    val productId: Long?,
+    val productId: Int?,
     val asin: String,
     val brand: String,
     val category: String,
