+++ server/src/main/kotlin/it/polito/wa2/g29/server/model/Attachment.kt
@@ -11,4 +11,4 @@ class Attachment(
     var type: AttachmentType,
     @ManyToOne
     var message: Message
-) : EntityBase<Long>()
+) : EntityBase<Int>()
