+++ server/src/main/kotlin/it/polito/wa2/g29/server/model/Chat.kt
@@ -8,7 +8,7 @@ class Chat(
     @OneToOne
     @MapsId
     val ticket: Ticket
-) : EntityBase<Long>() {
+) : EntityBase<Int>() {
     @OneToMany(mappedBy = "chat")
     var messages: Set<Message> = mutableSetOf()
 }
