+++ server/src/main/kotlin/it/polito/wa2/g29/server/model/Product.kt
@@ -13,4 +13,4 @@ class Product(
     var name: String,
     var price: Float,
     var weight: Float,
-) : EntityBase<Long>()
+) : EntityBase<Int>()
